Make NodeMapping an interface
